refactor: 기본정보 입력 훅으로 분리
